Refactored widgets to own module
